+++ docs/changelog.md
@@ -195,7 +195,7 @@
     - [x] 2.highight entity classes
         - enable if er diagram is enabled
         - entities in er diagram should be labeled as "Entity" after the title, and title should be bold
-    - [ ] 3.click esc to cancel search
+    - [x] 3.click esc to cancel search
     - [ ] 4.show loader name
     - [ ] 5.show relationship list when double click entity in er diagram
     - [ ] 6.highlight entity in use case